LPD-86298 Replacing poshi tests to playwright
